revolution!

+++ delta/src/main/java/org/byteam/delta/Delta.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 import android.util.Log;
 
 import java.io.File;
+import java.io.FileFilter;
 import java.io.IOException;
 
 
@@ -19,19 +20,42 @@ public class Delta {
     private Delta() {
     }
 
+    public static void applyPatchForTest(Context context, PatchListener patchListener) {
+        if (!isDebug(context)) {
+            Log.w(TAG, "Running test code in release build type!");
+        }
+        File[] patchs = Paths.getTmpDir(context).listFiles(new FileFilter() {
+            @Override
+            public boolean accept(File file) {
+                return file.isFile() && file.getName().endsWith(".apk");
+            }
+        });
+        if (patchs == null || patchs.length < 1) {
+            patchListener.patchResult(false, "Can not find patch file in tmp dir.");
+            return;
+        }
+        applyPatch(context, patchs[0], patchListener);
+    }
+
     public static void applyPatch(Context context, File apkFile, PatchListener patchListener) {
         applyPatch(context, apkFile, true, patchListener);
     }
 
-    public static void applyPatch(Context context, File apkFile, boolean isDiff, PatchListener patchListener) {
+    public static void applyPatch(Context context, File apkFile, boolean isPatch, PatchListener patchListener) {
         if (context == null) {
             throw new NullPointerException("context can not be null.");
         }
         if (apkFile == null) {
             throw new NullPointerException("apkFile can not be null.");
         }
-        if (!apkFile.exists() || apkFile.isDirectory() || apkFile.canRead()) {
-            throw new IllegalArgumentException("Invalid apk: " + apkFile.getPath());
+        if (!apkFile.exists()) {
+            throw new IllegalArgumentException("Invalid apk: " + apkFile.getPath() + " not exists.");
+        }
+        if (apkFile.isDirectory()) {
+            throw new IllegalArgumentException("Invalid apk: " + apkFile.getPath() + "is a directory.");
+        }
+        if (!apkFile.canRead()) {
+            throw new IllegalArgumentException("Invalid apk: " + apkFile.getPath() + " can't read.");
         }
         if (patchListener == null) {
             throw new NullPointerException("You must offer a 'PatchListener'.");
@@ -39,7 +63,7 @@ public class Delta {
 
         mPatchListener = patchListener;
 
-        if (isDiff) {
+        if (isPatch) {
             combineApk(context, apkFile);
         } else {
             try {
@@ -71,7 +95,7 @@ public class Delta {
             }
         }
         try {
-            FileUtils.copyFile(new File(apk), baseApkDir);
+            FileUtils.copyFileToDirectory(new File(apk), baseApkDir);
         } catch (IOException e) {
             Log.e(TAG, "Failure while extracting base.apk.", e);
             if (mPatchListener != null) {
@@ -135,12 +159,36 @@ public class Delta {
      */
     public static void clear(Context context) {
         try {
-            FileUtils.cleanDirectory(Paths.getDexDirectory(context));
-            FileUtils.cleanDirectory(Paths.getBaseApkDirectory(context));
-            FileUtils.cleanDirectory(Paths.getNewApkDirectory(context));
+            if (Paths.getDexDirectory(context).exists()) {
+                FileUtils.cleanDirectory(Paths.getDexDirectory(context));
+            }
+            if (Paths.getBaseApkDirectory(context).exists()) {
+                FileUtils.cleanDirectory(Paths.getBaseApkDirectory(context));
+            }
+            if (Paths.getNewApkDirectory(context).exists()) {
+                FileUtils.cleanDirectory(Paths.getNewApkDirectory(context));
+            }
+            if (Paths.getTmpDir(context).exists()) {
+                FileUtils.cleanDirectory(Paths.getTmpDir(context));
+            }
         } catch (IOException e) {
             Log.e(TAG, "Failed to clean patch file.", e);
         }
+    }
+
+    private static boolean isDebug(Context context) {
+        try {
+            Class<?> clazz = Class.forName(context.getPackageName() + ".BuildConfig");
+            Object debug = clazz.getDeclaredField("DEBUG").get(null);
+            return (boolean) debug;
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (NoSuchFieldException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+        return false;
     }
 
 }
